Move form creation to index file
